requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
distlib==0.3.9
filelock==3.18.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
platformdirs==4.3.8
PyMySQL==1.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
virtualenv==20.31.2
Werkzeug==3.1.3
